requirements.txt
annotated-types==0.7.0
anyio==4.9.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
dnspython==2.7.0
dulwich==0.22.8
fastapi==0.116.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
keyring==25.6.0
more-itertools==10.7.0
motor==3.7.1
msgpack==1.1.1
nodeenv==1.9.1
packaging==25.0
pbs-installer==2025.7.2
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
pre_commit==4.2.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.13.2
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.4
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomlkit==0.13.3
trove-classifiers==2025.5.9.12
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.31.2
zstandard==0.23.0
